move listProcessProperties into processes module
